+++ Runtime/Native/AffiseApiMethod.cs
@@ -31,8 +31,8 @@ namespace AffiseAttributionLib.Native
         // callbacks
         GET_REFERRER_URL_CALLBACK,
         GET_REFERRER_URL_VALUE_CALLBACK,
-        GET_REFERRER_ON_SERVER_CALLBACK,
-        GET_REFERRER_ON_SERVER_VALUE_CALLBACK,
+        GET_DEFERRED_DEEPLINK_CALLBACK,
+        GET_DEFERRED_DEEPLINK_VALUE_CALLBACK,
         REGISTER_DEEPLINK_CALLBACK,
         SKAD_REGISTER_ERROR_CALLBACK,
         SKAD_POSTBACK_ERROR_CALLBACK,
@@ -98,8 +98,8 @@ namespace AffiseAttributionLib.Native
                 // callbacks
                 AffiseApiMethod.GET_REFERRER_URL_CALLBACK => "get_referrer_url_callback",
                 AffiseApiMethod.GET_REFERRER_URL_VALUE_CALLBACK => "get_referrer_url_value_callback",
-                AffiseApiMethod.GET_REFERRER_ON_SERVER_CALLBACK => "get_referrer_on_server_callback",
-                AffiseApiMethod.GET_REFERRER_ON_SERVER_VALUE_CALLBACK => "get_referrer_on_server_value_callback",
+                AffiseApiMethod.GET_DEFERRED_DEEPLINK_CALLBACK => "get_deferred_deeplink_callback",
+                AffiseApiMethod.GET_DEFERRED_DEEPLINK_VALUE_CALLBACK => "get_deferred_deeplink_value_callback",
                 AffiseApiMethod.REGISTER_DEEPLINK_CALLBACK => "register_deeplink_callback",
                 AffiseApiMethod.SKAD_REGISTER_ERROR_CALLBACK => "skad_register_error_callback",
                 AffiseApiMethod.SKAD_POSTBACK_ERROR_CALLBACK => "skad_postback_error_callback",