+++ Runtime/Parameters/Providers/AffSDKVersionProvider.cs
@@ -9,6 +9,6 @@ namespace AffiseAttributionLib.AffiseParameters.Providers
     {
         public override float Order => 47.0f;
         public override ProviderType? Key => ProviderType.AFFISE_SDK_VERSION;
-        public override string Provide() => "1.6.32";
+        public override string Provide() => "1.6.33";
     }
 }